Move all fallback safety logic to FallbackHandlingResourceHasher
